+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 
+- Add support for extensions using `ExtType`.
 - Add the Python implementation of `mashpack.Packer`.
 - Add the Python implementation of `mashpack.Unpacker`.
 - Add the `ARRAY8`, `ARRAY16`, `ARRAY32` data types for typed arrays to save